requirements.txt
natsort>=8.4.0
Pillow>=11.2.1
python_pptx>=1.0.2
